CardAssignController add authorize

+++ app/Http/Controllers/Api/CardAssignController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\Card\User\DestroyCardUserRequest;
 use App\Http\Requests\Card\User\StoreCardUserRequest;
 use App\Http\Resources\Card\CardAssignedUserResource;
 use App\Models\Card;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 
 class CardAssignController extends Controller
@@ -14,9 +15,11 @@ class CardAssignController extends Controller
     /**
      * @param Card $card
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function index(Card $card): JsonResponse
     {
+        $this->authorize('view', $card->board());
         return $this->successResponse(CardAssignedUserResource::collection($card->assignedUsers), 'Success');
     }
 
@@ -24,9 +27,11 @@ class CardAssignController extends Controller
      * @param StoreCardUserRequest $request
      * @param Card $card
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function update(StoreCardUserRequest $request, Card $card): JsonResponse
     {
+        $this->authorize('update', $card->board());
         $card->assignedUsers()->attach($request->users);
         return $this->successResponse(CardAssignedUserResource::collection($card->assignedUsers), 'success');
     }
@@ -35,9 +40,11 @@ class CardAssignController extends Controller
      * @param DestroyCardUserRequest $request
      * @param Card $card
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function destroy(DestroyCardUserRequest $request, Card $card): JsonResponse
     {
+        $this->authorize('delete', $card->board());
         $card->assignedUsers()->detach($request->users);
         return $this->successResponse(CardAssignedUserResource::collection($card->assignedUsers), 'Success');
     }
